Add localization to templates and adapt templates to event translations
The templates at first had hard coded strings in them which were
replaced by translatable strings. The templates also were adapted to
the seperation of EventForm and EventTranslationForm. Furthermore,
filters for usage in the templates were introduced which were just
copied from the page filters and adapted to the events. This should be
implemented more generic in the future.